02.simple_rag_langchain_openai/app/requirements.txt
langchain
langchain-openai
langchain-chroma
python-dotenv
fastapi
uvicorn